requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_login
flask_wtf
wtforms
flask_mail
requests